Logical decoding - the minimal functionality.
This commit implements the actual decoding of the zheap XLOG records, so that
simple use cases work. I chose to store the data in the zheap tuple format
until it's processed. The tuple is converted to the traditional heap tuple
before it's passed to the output plugin, so the output plugin sees no
difference.

Serialization (mis)uses the fact that ZHeapTupleData structure currently has
the same layout as HeapTupleData. (StaticAssertStmt is used to keep eye on
this.)

There are at least 2 problems left:

1. zheap tuple does not contain the command id (CID), so we will need to
retrieve it from the UNDO log for the logical decoding to use the correct
historic snapshots.

2. zheap always sets xl_xid of XLogRecord to the top transaction, so
reorderbuffer.c cannot discard aborted subtransactions.

The best idea I have so far is to let ApplyUndoActions() insert a special XLOG
record just before it returns (and when wal_level >= logical) and wrap the XID of the aborted subtransaction into it.