Added subtransaction XID to the insert/update/delete XLOG records.
This is to fix the subtransaction problem reported in the commit message of
95343f30. Alternatively we could retrieve the XID from the UNDO log, but that
would make it necessary to postpone discarding of the UNDO log in a similar
way the replication slots preserve WAL - such an approach seems more
difficult.

Before we consider the logical decoding complete, we should adjust the test
plugin (contrib/test_decoding) so that it generates the same output as it does
for the heap AM. The problem is that zheap does provide the plugin with the
"old tuple" in cases where heap does not.

While adjusting the zheap REDO functions, I've noticed some misaligned access
to the XLOG records, so fixed those.

The 95343f30 commit message also mentions command id, but that's really not
needed for zheap.