Flatten the "old-tuple" for logical decoding before inserting it into WAL.
The tuple must not contain external attributes because the TOAST tuple (or
even table) might be gone at decoding time. The same flattening actually
happens for the heap AM, however for zheap we need to process all the
attributes, not only the identity key - the reason is that the tuple might be
needed for recovery.