Write the TOASTed identity attributes to WAL.
With the heap AM, the old tuple is flattened (i.e. TOASTed attributes are
inlined) and written to WAL so that it can be logically decoded
easily. However zheap may need the old tuple also for REDO (recovery), so no
flattening is acceptable (i.e. exactly the same tuple that we replace in the
table, including TOAST pointers, must be WAL-logged). However we still need
the values of the TOASTed attributes to be WAL-logged because the TOAST values
are not guaranteed to exist during the logical decoding.

With this commit, logical decoding causes values of the TOASTed attributes
participating in the identity key to be WAL-logged. Before particular data
change (UPDATE / DELETE) is replayed, we create a copy of the tuple that would
normally be used for REDO, add (inline) those values of identity attributes to
it and pass the tuple to the output plugin.

It'd be simpler to write the flattened old tuple to WAL in addition to the
non-flattened one, but it'd obviously be wasted WAL space.